requirements.txt
streamlit
pdfplumber
pillow
pytesseract
sentence-transformers
faiss-cpu
langchain_groq
transformers
torch
tqdm